+++ src/flows/Onboarding/components/OnboardingForm.tsx
@@ -1,7 +1,7 @@
 import { JSONSchemaFormFields } from '@/src/components/form/JSONSchemaForm';
 import { Form } from '@/src/components/ui/form';
 import { useEffect } from 'react';
-import { useOnboardingContext } from './context';
+import { useOnboardingContext } from '@/src/flows/Onboarding/context';
 import { useForm } from 'react-hook-form';
 import { useJsonSchemasValidationFormResolver } from '@/src/components/form/yupValidationResolver';
 import { Fields } from '@remoteoss/json-schema-form';
