+++ src/flows/Onboarding/tests/OnboardingSubmit.test.tsx
@@ -2,7 +2,7 @@
 import { useFormFields } from '@/src/context';
 import { useOnboardingContext } from '@/src/flows/Onboarding/context';
 import { fireEvent, render, screen } from '@testing-library/react';
-import { OnboardingSubmit } from '../OnboardingSubmit';
+import { OnboardingSubmit } from '../components/OnboardingSubmit';
 
 // Mock dependencies
 vi.mock('@/src/context', () => ({
